Refactor integration tests to eliminate duplication and fix import styles
Major refactoring of integration tests to follow DRY principles and improve maintainability:

## Integration Test Refactoring:
- Remove redundant test_sonarr_container_integration (overlapped with scanner-only test)
- Replace 4 separate test functions with single parameterized test covering all scenarios
- Create reusable test_helpers.py module with 375 lines of utility functions
- Add prepared_series_with_nfos fixture for common setup/teardown
- Reduce main test file from 829 lines to 160 lines (80% reduction)
- Eliminate 90%+ code duplication across test cases

## Import Style Fixes:
- Move all imports to top of files per Python conventions
- Fix subprocess_service_manager.py: move shutil import from method to top
- Fix conftest.py: move test_helpers import to top-level imports
- Fix test_sonarr_integration.py: consolidate all helper imports at top
- Fix test_metadata_processor.py: remove 4 redundant inline TranslatedContent imports
- Fix test_rewrite_service.py: move TranslatedContent import to top

## Benefits:
- Much easier to maintain and debug test failures
- Consistent test behavior across all integration scenarios
- Simple to add new service configurations via parameters
- Cleaner code following Python import conventions
- All tests still pass with identical coverage
